refactor: remove dashboardService in favour of publicationService
This change consolidates the `dashboardService.getMyPublicationsByTopic`
into `publicationService` as this makes more sense grouping wise.
