+++ src/explore-education-statistics-admin/src/services/publicationService.ts
@@ -1,13 +1,31 @@
 import { ContactDetails, SaveContact } from '@admin/services/contactService';
-import { ExternalMethodology } from '@admin/services/dashboardService';
 import {
   LegacyRelease,
   UpdateLegacyRelease,
 } from '@admin/services/legacyReleaseService';
 import { BasicMethodology } from '@admin/services/methodologyService';
+import { MyRelease } from '@admin/services/releaseService';
 import { IdTitlePair } from '@admin/services/types/common';
 import client from '@admin/services/utils/service';
 import { OmitStrict } from '@common/types';
+
+export interface ExternalMethodology {
+  title: string;
+  url: string;
+}
+
+export interface MyPublication {
+  id: string;
+  title: string;
+  methodology?: IdTitlePair;
+  externalMethodology?: ExternalMethodology;
+  releases: MyRelease[];
+  contact?: ContactDetails;
+  permissions: {
+    canCreateReleases: boolean;
+    canUpdatePublication: boolean;
+  };
+}
 
 export interface BasicPublicationDetails {
   id: string;
@@ -41,6 +59,11 @@ export type UpdatePublicationLegacyRelease = Partial<
 >;
 
 const publicationService = {
+  getMyPublicationsByTopic(topicId: string): Promise<MyPublication[]> {
+    return client.get<MyPublication[]>('/me/publications', {
+      params: { topicId },
+    });
+  },
   createPublication(
     publication: CreatePublicationRequest,
   ): Promise<BasicPublicationDetails> {
